xnetcommunities2wordcloud/requirements.txt
python-igraph
matplotlib
numpy
wordcloud
